+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.3.0 ###
 * Added admin settings page for the plugin.
+* Added admin option for featured post image in RSS feeds.
 
 ### 2.2.0 ###
 * Display featured post image in RSS feeds.